requirements.txt
# AP 原型最小运行依赖（建议）
# ---------------------------------
# 说明：
# - 当前原型大部分模块只依赖 Python 标准库；
# - 但 YAML 配置读取与中文分词属于“强烈推荐”能力，公开到 GitHub 时建议明确列出。

PyYAML>=6.0
jieba>=0.42.1
